backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.4.0
django-redis==6.0.0
djangorestframework==3.15.2
fastapi==0.116.1
gunicorn==21.2.0
h11==0.16.0
idna==3.10
joblib==1.5.2
numpy==2.3.2
packaging==25.0
pandas==2.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==5.0.7
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.3
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
